Add KeyExpiryDisabled field to Device
This commit adds the `KeyExpiryDisabled` boolean field to the `Device` type. This field indicates that a device
key has been configured to never expire.

This field is required for https://github.com/davidsbond/terraform-provider-tailscale/issues/78 on the terraform
provider so that we can perform a read on the device to determine the existing properties of a device key.

Signed-off-by: David Bond <davidsbond93@gmail.com>