Do not attempt to standardise response bodies when we don't need to
Related to https://github.com/davidsbond/terraform-provider-tailscale/issues/113

This commit modifies the request making code to not attempt to standardise
response bodies if we are not attempting to parse the response. For example, the
tailnet key delete API response has an empty body, which would cause errors when
calling hujson.Standardize.

Signed-off-by: David Bond <davidsbond93@gmail.com>