Add method for ACL validation
This commit adds a `ValidateACL` method to the `Client` type that will be used to check the
ACL is well formatted and the specified tests pass. It uses the JSON object method described
in the API documentation.

This will be required for implementing additional validation as requested on the terraform
provider:

https://github.com/tailscale/terraform-provider-tailscale/issues/158

Signed-off-by: David Bond <davidsbond93@gmail.com>